requirements.txt
python-telegram-bot==20.7
asyncpg==0.29.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6